make wicket-threadtest tests
