+++ testing/wicket-threadtest/src/test/java/org/apache/wicket/threadtest/App1Test.java
@@ -17,22 +17,61 @@
 package org.apache.wicket.threadtest;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import org.apache.wicket.threadtest.apps.app1.ResourceTestPage;
 import org.apache.wicket.threadtest.tester.SimpleGetCommand;
 import org.apache.wicket.threadtest.tester.Tester;
+import org.junit.Test;
 
 /**
- * @author almaw
+ * @author eelcohillenius
  */
-public class App1Test3
+public class App1Test
 {
-	/**
-	 * @param args
-	 * @throws Exception
-	 */
-	public static void main(String[] args) throws Exception
+
+	@Test
+	public void test1() throws Exception
+	{
+
+		List<String> gets = Arrays.asList(
+			"/app1/wicket/bookmarkable/org.apache.wicket.threadtest.apps.app1.Home",
+			"/app1/wicket/page?${id}-1.-link");
+
+		// you can turn this on if you e.g. want to attach to a profiler
+// Thread.sleep(5000);
+
+		SimpleGetCommand getCmd = new SimpleGetCommand(gets, 10) {
+			@Override
+			protected String getValue(String name, int iteration) {
+				if ("id".equals(name)) {
+					// page id increases by one on each render
+					return "" + (iteration * 2);
+				}
+				return super.getValue(name, iteration);
+			}
+		};
+		// getCmd.setPrintResponse(true);
+		Tester tester = new Tester(getCmd, 100, true);
+		tester.run();
+	}
+	
+	@Test
+	public void test2() throws Exception
+	{
+
+		List<String> gets = Arrays.asList("/app1/wicket/bookmarkable/org.apache.wicket.threadtest.apps.app1.Home");
+
+		SimpleGetCommand getCmd = new SimpleGetCommand(gets, 5);
+
+		// getCmd.setPrintResponse(true);
+		Tester tester = new Tester(getCmd, 100, false);
+		tester.run();
+	}
+	
+	@Test
+	public void test3() throws Exception
 	{
 
 		List<String> gets = new ArrayList<String>();
